requirements.txt
flask
pandas
scikit-learn
pytesseract
Pillow
